cast on the uint64_t to an int64_t 64 bit should never be set

+++ include/Opensea_Parser_Helper.h
@@ -37,6 +37,7 @@ extern char *g_currentTimeStringPtr;
 
 
 namespace opensea_parser {
+
 #ifndef OPENSEA_PARSER
 #define OPENSEA_PARSER
 
@@ -228,7 +229,7 @@ namespace opensea_parser {
 		}
 		else
 		{
-			if (value < INT32_MAX && value > INT32_MIN)
+			if ((int64_t)value < INT32_MAX && (int64_t)value > INT32_MIN)
 			{
 				json_push_back(nowNode, json_new_i((char *)myStr.c_str(), static_cast<int32_t>(M_DoubleWord0(value))));
 			}
